Almost complete rewrite of the whole program. The structure has changed to be more ordered, faster, and reliable especially to exceptional conditions, where set names are named like actions, options, file names, or are named `-' (a hyphen), or start with one.
Now works with set names of add/del options (i.e. timeout,before,...).

If adding elements do not display the 'timeout' option,
if the set was not created with timeout support.

Retrieval and deduping of options is a function now
and is only called as needed.

If the bash_completion package functions are available,
the colon is no longer removed from COMP_WORDBREAKS.

If the functions of the bash_completion package are available,
those are used, otherwise a workaround is used. Though
bash_completion 2.0+ is recommended.